services/ddf_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic>=2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart>=0.0.6
mcp>=1.0.0,<2.0.0

